requirements.txt
streamlit
openai
requests
smtplib  # (via standard library; listed for clarity)
schedule
pandas
google-api-python-client
google-auth
google-auth-oauthlib
oauth2client
plotly
python-dotenv
